bootstrap: move uv to it's own module
We move the handling of the UV installation to it's own module,
so that we can reuse it for when we currently use the installed
uv into self.
