+++ rye/src/main.rs
@@ -18,6 +18,7 @@ mod pyproject;
 mod sources;
 mod sync;
 mod utils;
+mod uv;
 
 static SHOW_CONTINUE_PROMPT: AtomicBool = AtomicBool::new(false);
 static DISABLE_CTRLC_HANDLER: AtomicBool = AtomicBool::new(false);
